requirements.txt
streamlit
langchain
langchain_community
pypdf
python-dotenv
faiss-cpu